+++ CMakeLists.txt
@@ -173,9 +173,12 @@ set(SPARROW_SRC
     ${SPARROW_SOURCE_DIR}/array.cpp
     ${SPARROW_SOURCE_DIR}/array_factory.cpp
     ${SPARROW_SOURCE_DIR}/array_helper.cpp
+    ${SPARROW_SOURCE_DIR}/arrow_array_schema_proxy.cpp
+    ${SPARROW_SOURCE_DIR}/arrow_interface/arrow_array.cpp
+    ${SPARROW_SOURCE_DIR}/arrow_interface/arrow_schema.cpp
     ${SPARROW_SOURCE_DIR}/list_value.cpp
-    ${SPARROW_SOURCE_DIR}/struct_value.cpp
     ${SPARROW_SOURCE_DIR}/run_encoded_array.cpp
+    ${SPARROW_SOURCE_DIR}/struct_value.cpp
 )
 
 add_library(sparrow SHARED ${SPARROW_HEADERS} ${SPARROW_SRC})
