+++ src/arrow_interface/arrow_array.cpp
@@ -1,0 +1,110 @@
+// Copyright 2024 Man Group Operations Limited
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+#include "sparrow/arrow_interface/arrow_array.hpp"
+
+#include "sparrow/arrow_interface/arrow_array_schema_info_utils.hpp"
+#include "sparrow/types/data_type.hpp"
+
+
+namespace sparrow
+{
+    void release_arrow_array(ArrowArray* array)
+    {
+        SPARROW_ASSERT_FALSE(array == nullptr)
+        SPARROW_ASSERT_TRUE(array->release == std::addressof(release_arrow_array))
+
+        if (array->private_data != nullptr)
+        {
+            const auto private_data = static_cast<arrow_array_private_data*>(array->private_data);
+            delete private_data;
+            array->private_data = nullptr;
+        }
+        array->buffers = nullptr;  // The buffers were deleted with the private data
+        release_common_arrow(*array);
+    }
+
+    std::vector<sparrow::buffer_view<uint8_t>>
+    get_arrow_array_buffers(const ArrowArray& array, const ArrowSchema& schema)
+    {
+        std::vector<sparrow::buffer_view<uint8_t>> buffers;
+        const auto buffer_count = static_cast<size_t>(array.n_buffers);
+        buffers.reserve(buffer_count);
+        const enum data_type data_type = format_to_data_type(schema.format);
+        const auto buffers_type = get_buffer_types_from_data_type(data_type);
+        for (std::size_t i = 0; i < buffer_count; ++i)
+        {
+            const auto buffer_type = buffers_type[i];
+            auto buffer = array.buffers[i];
+            const std::size_t buffer_size = compute_buffer_size(
+                buffer_type,
+                static_cast<size_t>(array.length),
+                static_cast<size_t>(array.offset),
+                data_type,
+                buffers,
+                i == 0 ? buffer_type : buffers_type[i - 1]
+            );
+            auto* ptr = static_cast<uint8_t*>(const_cast<void*>(buffer));
+            buffers.emplace_back(ptr, buffer_size);
+        }
+        return buffers;
+    }
+
+    void copy_array(const ArrowArray& source_array, const ArrowSchema& source_schema, ArrowArray& target)
+    {
+        SPARROW_ASSERT_TRUE(&source_array != &target);
+        SPARROW_ASSERT_TRUE(source_array.release != nullptr);
+        SPARROW_ASSERT_TRUE(source_schema.release != nullptr);
+        SPARROW_ASSERT_TRUE(source_array.n_children == source_schema.n_children);
+        SPARROW_ASSERT_TRUE((source_array.dictionary == nullptr) == (source_schema.dictionary == nullptr));
+
+        target.n_children = source_array.n_children;
+        if (source_array.n_children > 0)
+        {
+            target.children = new ArrowArray*[static_cast<std::size_t>(source_array.n_children)];
+            for (int64_t i = 0; i < source_array.n_children; ++i)
+            {
+                SPARROW_ASSERT_TRUE(source_array.children[i] != nullptr);
+                target.children[i] = new ArrowArray{};
+                copy_array(*source_array.children[i], *source_schema.children[i], *target.children[i]);
+            }
+        }
+
+        if (source_array.dictionary != nullptr)
+        {
+            target.dictionary = new ArrowArray{};
+            copy_array(*source_array.dictionary, *source_schema.dictionary, *target.dictionary);
+        }
+
+        target.length = source_array.length;
+        target.null_count = source_array.null_count;
+        target.offset = source_array.offset;
+        target.n_buffers = source_array.n_buffers;
+
+
+        std::vector<buffer<std::uint8_t>> buffers_copy;
+        buffers_copy.reserve(static_cast<std::size_t>(source_array.n_buffers));
+        const auto buffers = get_arrow_array_buffers(source_array, source_schema);
+        for (const auto& buffer : buffers)
+        {
+            buffers_copy.emplace_back(buffer);
+        }
+        target.private_data = new arrow_array_private_data(std::move(buffers_copy));
+        const auto private_data = static_cast<arrow_array_private_data*>(target.private_data);
+        target.buffers = private_data->buffers_ptrs<void>();
+        target.release = release_arrow_array;
+    }
+
+
+}
